[MRESOLVER-160] Deprecate service locator
Deprecate service locator implementation as well, as integrators
usually extend this class (see Maven for example), and they don't
deal with deprecated interfaces, so the fact those are deprecated
as may be missed at first sight.

By deprecating this implementation class as well, the integrators
will get a warning in their build and/or IDEs, so intent will be
more obvious.