Introduce ResolverLock, layer on top of low level Named Locks
The resolver lock adds a thin layer above low lever named locks,
does not leak actualy named lock names, but is keyed by artifact,
hence lock ordering is stable (is based on Artifact not on name mapper).
Finally, resolver lock factory is able to spice up the logic for locking
that is now "hacked in" to not alter SyncContextFactory API.

This is EXPERIMENT for testing reasons only.