[MRESOLVER-253] Split LRM (#168)
Enhances existing "enhanced" LRM (local repository manager) with "prefix" from from composer. Default behaviour is not changed (will not use prefix, local repo will be as today).

The point in change is that it introduces a "composer", that is composing LRM path prefix, and is able to apply different strategies, and split local repository into "installed" (locally built and installed) and "cached" (downloaded from remote), etc.

There are several composers options out of the box.