[MRESOLVER-265] Checksum of checksum
There is a discrepancy when a checksum is attached (so not added
by resolver itself), and resolver will checksum it if the algorithm
is not configured (ie. default config used but SHA512 added).

This makes it work as expected, as checksum recognition works
now against ALL SUPPORTED and not configured ONLY checksums).