[MRESOLVER-266] Simplify named lock adapter creation and align configuration
Simplified adapter creation and now as everything else in Resolver, session
config properties are used as configuration source, not Java system
properties.

Important: This change also drops internal/impl things. Therefore, this is a
breakage, IF someone uses internal/private/impl details from Resolver
(mvnd does, but is fine, is our kiddo), and for the rest of the worlds:
hands off, please.

This closes #196