[MRESOLVER-291] Undo params use (#219)
Pull out ctor param injection, provide generic enough helper to allow anyone to customize as they want.

Changes in short: 
Nor SyncContextFactory nor NamedLockFactoryAdapter should shut down. This was the problem, and I was mislead by it. NamedLockFactory instances are **singletons**, and **are inited in ctor**. Hence, if two adapters used same factory, in case one adapter is shut down, it would shut down the lock factory of the other one as well (unexpectedly). Hence, it is solely the adapter factory duty to shut down factories, but again, as we saw, they are inited in ctor, hence if we assume there is living HZ named lock factory, even if unused (as file lock is used for example), shutdown is still needed, to shut down the HZ instance (same applies to Redisson btw)! All in all, this new class is IMHO the most correct, while it allows overriding and changing things, if needed.

Minor: do not create static initializers to be used in SL, make them methods invoked when needed.

---

https://issues.apache.org/jira/browse/MRESOLVER-291