Unjustified use of collocated temp file
What happens here is validator using transporter to download
external checksum to temp file, only to immediately load it up
into memory (into checksumExpectedValues map) and dispose the
tmp file.

Still, here a collocated temp file was used, which was wrong,
as the temp file is not "moved" to it's final place, it is only
used as target for transport to download the checksum file.

Hence, it is not the local repo that needs to be (potentially)
polluted, but /tmp is just fine for this.