[] Make javax.inject optional
javax.inject is and should be optional (used only when Guice and/or
Sisu is being used), but in case SL is being used no javax.inject
should be needed.

This rule was violated in 1.9.0 commit 5566bd5b3b0e59d124d820e6274da8e2da7804b0
where javax.inject.Provider was used in conjuction with SL as well.

---

https://issues.apache.org/jira/browse/MRESOLVER-298