[MRESOLVER-320] Return to original pool setup
There was some information lost on the way how we got to current
master: artifact and dependency was ALWAYS weak, and the "cause"
commit 6dda69980cf4c13228a4f6561bfa0d8384a0be68 modified ONLY
the descriptor pool (was hard by mistake, made it weak). This also
explain why we saw with yourkit excess Xpp3 and model building
requests as well (as starting with 1.8.x descriptors/POMs were
GCed, while before they were not).

This PR essentially returns the state of affairs to pre 1.8.x
resolver state re data pools, but also adds ability to "tune" each
three of them separately.

---

https://issues.apache.org/jira/browse/MRESOLVER-320