[MRESOLVER-320] Return to original pool setup (#243)
The original pool setup was always (weak, weak, hard), but by mistake. The resolver 1.8.0 got a fix for reported MRESOLVER-250: (that did lower the memory consumption but added 10% execution time overhead). The fix for MRESOLVER-250 caused that GC collect descriptors, so during collection models were continuously been rebuilt, but this operation is all but "cheap" (a lot of CPU and a lot of IO).

This commit rewrites the pools, makes them configurable (for artifacts, dependencies and descriptors/POMs) . With this PR now pools are configurable as well (as hard+hard+hard setup may offer even speedup at cost of more heap usage).

It also FIXES a bug creeped into master (was never released) where interning pool misbehaves: it was able to return null  value.

---

https://issues.apache.org/jira/browse/MRESOLVER-320