[MRESOLVER-376] Bugfix for relocated artifacts
The code uses Artifact to key (forever during collection) the
read artifact descriptors, but, this is not always what we want.

Example: ArtifactDescriptorReader when reads a relocated artifact,
it will use RelocatedArtifact (in maven-resolver-provider) as
result.artifact, and BF use this artifact as key. But, this artifact
will be key of relocation target, while the descriptor is the
descriptor of original (relocated) artifact. As it gets into
(inner) async resolver cache, it will stay there forever.
But alas, we next ask for relocation target decriptor, that
as can be seen, use same (key-wise, or GAV wise) key as
the relocated artifact, and nothing happens, as computeIfAbsent
does not compute (as it is present), and we fall into
endless loop.

---

https://issues.apache.org/jira/browse/MRESOLVER-376