[MRESOLVER-376] UT for bug
This UT on master makes old DF work, while BF break with SOEx.

If same UT ran on fixed branch, both pass.

Still, seems combo of range+relocation has some other issues as well,
as all this works only partially, as the range is not resolved,
only first applicant (lowest version in DF and highest version
in BF as it sort-reverse the versions).

---

https://issues.apache.org/jira/browse/MRESOLVER-376