[MRESOLVER-586] Avoid copying the type default properties map
When we don't have specific properties for a DefaultArtifact, we can
just reuse the type default properties map as it's readonly and
immutable.
The patch contains some additional changes as I renamed the method
and parameters to make them more specific so that the method
wouldn't be used for things that wouldn't be safe.