[MRESOLVER-605] Linking support in file transporter (#578)
Support linking for FileTransport. Still, as FileTransport is remote transport, due checksumming (via Listener) file does have to be read.

Big fat note: this change will cause that symlinks may land in local repository. This PR does NOT enforce rest of Maven plugins knows how to handle them, is out of scope for this PR.

This feature is EXPERIMENTAL.

---

https://issues.apache.org/jira/browse/MRESOLVER-605