[MRESOLVER-569] Pass on exceptions (#678)
Pass on received exceptions as suppressed. Resolver was conceived when no suppressed exceptions existed and does not make use of them at all. This PR leaves existing "cause" management intact, but passes the whole tree (even for notorious `ArtifactResolutionException`) the full tree as suppressed exceptions. 

This implies that old code can still make use of existing logic to create the (incomplete) error messages, while new code can get all the needed information by walking and including suppressed exceptions as well. Logic is: if there is suppressed (and is same instance as cause or cause is null), then the given cause, if any, is "just the first out of many suppressed", hence suppressed "prevails" the cause. In that case code needs to "follow" suppressed ones.

Example: created a POM with non existent dependency, and added 3 extra reposes (so 4 in total with Central):
https://gist.github.com/cstamas/2d770f073a0260debb68dcc7ab3a7799
The whole history is now in exceptions.

---

https://issues.apache.org/jira/browse/MRESOLVER-569