Ability to augment metadata nature for version range request
The idea is to let caller be able to augment the nature of
metadata to resolve version range from.

Resolver already handles this, and is more elegant, than
to have caller fiddle with filtering the remote repositories
or alike. Also, this way client code (ie. Maven) can expose
more fine grained configuration for this.