ConflictResolver improvements, preparations (#1542)
Improvements to conflict resolver, step one: cleanup and return type safety. This codebase is last that is 100% equal between Resolver 1 and Resolver 2, and was quite "dirty" at least what type safety matters. On may places `Object` was used as type. This PR merely cleans this up, and also revealed one interesting fast: at runtime Maven used `Integer` as conflictID while in tests `String` (for easier debug). This is now aligned, and `String` is used, but at runtime it is still integer counter String representation that is interned (to retain pooled numbers behavior). 

Minor improvements added to `Key` class that now memorizes `hashCode` instead to over and over recalculate it.

After this cleanup we should be able to more easily comprehend and improve conflict resolution.