Prefix file magic detection
Resolver now with enabled prefix file discovery discovers
even wrong files, most specifically in Maven ITs where
Jetty handler is not mounted to URL and receives all
sort of garbage, that is assumed to be prefix file.

From now on, prefix file must start with following
line (be the very first line in prefixes.txt):
```
```

And then and only then is recognized as prefixes file.