SmartExecutor: on classic pools TCCL must be delegated (#1553)
When Maven/Resolver runs on older than Java 21, the classic thread pools are used, and TCCL must be delegated now that they are reused across session, otherwise lookup issues occurs as Plexus DI will filter components based on it.

As Maven ITs shows, on Java 17 the DAV related ITs fail. They load the DAV Wagon as build extension (from POM). This means that BF used pool is instantiated way before, and pooled threads TCCL are not "aware" of (later) loaded up build extension (IT failes are "no DAV transport available", while the extension _is loaded up_).

With this change locally the ITs pass OK on Java 17, while on Java 21 (where VTs are used) this was never an issue, as we use "one VT per task" and VT is dropped once done (they are not pooled).

Before SmartExecutors, this was not a problem in BF collector, as pools were created "in situ" and then thrown away, that was sub-optimal.