Bug: Reproducer and fix for #1583 (#1585)
The new "path" CR in certain cases (cycle involving two different instances of `DependencyNode` as one was managed into same as other + parent of the cycle node has multiple/mixed children of cycling and non cycling + parent of the cycle head was the sole winner) left cycles in place even at verbosity levels of `NONE` and `STANDARD` when cycles should be removed. This is fixed. Another issue was that "path" CR was "too eager" by applying winner scope/optional to all nodes, while it should have apply it only to winner, as losers are marked as such (and possibly eliminated).

Changes:
* fixed the bug, also cleaned up `push(int)` method as it had some redundancy and misplaced comments
* fixed the eager application of winner scope/optional -- only winner have those applied
* dropped some copy-pasta constant remnants (they are all in super class)
* improved `Verbosity` javadoc re loops
* added UT that reproduces the issue covering all CRs and verbosity levels
* added demo with this exact case; demos also run as part of `run-its`, so is an extra safety belt

Fixes #1583