Proper fix
We already had means to "make sure string is valid for path segment",
but it was buried in RepositoryIdHelper.

Pull it out, and make it reusable, as this is essentially very
same use case. Also, the "expensive" method really punishes
ONLY non maven metadata, and those should be few only.