Stop proliferation of static helper IO methods (#1612)
Originally, the `PathProcessor` (before the now deprecated `FileProcessor`) was main component dealing with IO, restore this state.

The proliferation of static IO helpers is just bad, as it does not allow (easy) overriding as it is the case of components.

This fixes another pain point: the component interface `PathProcessor` (defined in SPI) has to have two (identical, almost) implementations: one in test-utils and one in impl. Stop this madness as well, now one "support" class is provided that implements everything and is published as component in SPI, but allows simple reusability in tests as well. And, same code is being tested finally.