Bug: fix prefix filter discovery
Two major issues: one was that it tried to fix the Maven issue
of remote repository uniqueness (globally), and that attempt
was in fact wrong, fix is elsewhere. The reason it tried to do
this is in fact to circumvent locking issues #1644.

The other issue was filter lifecycle, manager did it wrongly:
the bug caused that one session was used to acquire filters
and same filter got used for potentially other (maybe even
reconfigured) session. This overlook forced filters to
implement hoops and loops (to prevent recursion), but also
prevented any third party code to have saying in filter
operation.

Changes:
* fix manager to make sure filters and session are aligned
* fix/simplify prefix filter recursion prevention
* introduce public and documented way to inhibit prefix discovery